Refactor: Button (export default -> named export)
