+++ src/components/BottomSheet/BottomSheet.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Button from '~/components/Button/Button';
+import { Button } from '~/components/Button';
 
 import BottomSheet from './BottomSheet';
 import { useBottomSheet } from './useBottomSheet';
