+++ src/modules/IdCardCreation/IdCardCreationSteps.stories.tsx
@@ -1,6 +1,6 @@
 import type { StoryObj } from '@storybook/react';
 
-import Button from '~/components/Button/Button';
+import { Button } from '~/components/Button';
 import { ImageUrlMockHandler } from '~/mocks/image/image.mockHandler';
 import { IdCard } from '~/modules/IdCard/IdCard.client';
 import { BoardingStep, LoadingStep } from '~/modules/IdCardCreation/Step';
