+++ src/modules/IdCardEditButton/IdCardEditButton.client.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { usePathname, useRouter } from 'next/navigation';
 
-import Button from '~/components/Button/Button';
+import { Button } from '~/components/Button';
 
 export const IdCardEditButton = () => {
   const pathname = usePathname();
