+++ src/modules/PlanetEnterButton/PlanetEnterButton.client.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Button from '~/components/Button/Button';
+import { Button } from '~/components/Button';
 import TextButton from '~/components/Button/TextButton';
 
 export const PlanetEnterButton = () => {
